Witgen: Extract block production from Generator into VmProcessor
